+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Additional tabs "Albums" and "Tracks" to the artist details pane
 - Additional tabs "Tracks" and "Artists" to the album details pane
 - Favorite toggle to the details pane of the tracks, albums, artists, playlists, and podcasts
+- New filter "Favorite" for the smart list
 
 ### Changed
 - Drop support for PHP versions older 7.4 (i.e. PHP 7.1 - 7.3)